Release 5.0.0

+++ database/migrations/2022_02_24_124950_change_primary_discord_roles.php
@@ -4,15 +4,17 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class ChangePrimaryDiscordRoles extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
-     *
-     * @return void
      */
-    public function up()
+    public function up(): void
     {
+        if (Schema::getConnection()->getDriverName() === 'sqlite') {
+            return;
+        }
+
         Schema::table('discord_roles', function (Blueprint $table) {
             $table->dropColumn('aid');
             $table->primary('id');
@@ -21,11 +23,9 @@ class ChangePrimaryDiscordRoles extends Migration
 
     /**
      * Reverse the migrations.
-     *
-     * @return void
      */
-    public function down()
+    public function down(): void
     {
         //
     }
-}
+};
